CostOS logo as 'Loading' spinner background
A rough approximation of the Nomitech logo was made in MS Word with an image rotated via a macro, exported to PDF and then PNG. The resulting PNG images were cropped and had their background made transparent.

These were bundled together as a '.gif' file and used as a 'spinner' whilst waiting for a page to load it's data.